checkpoint - integrate-other controller first pass
